feat: extract event creation logic and ui
