+++ apps/epicdev-ai/src/app/(content)/events/[slug]/edit/_components/edit-event-form.tsx
@@ -2,7 +2,6 @@
 
 import * as React from 'react'
 import { useRouter } from 'next/navigation'
-import { DateTimePicker } from '@/app/(content)/events/[slug]/edit/_components/date-time-picker/date-time-picker'
 import { onEventSave } from '@/app/(content)/events/[slug]/edit/actions'
 import StandaloneVideoResourceUploaderAndViewer from '@/app/(content)/posts/_components/standalone-video-resource-uploader-and-viewer'
 import { TagField } from '@/app/(content)/posts/_components/tag-field'
@@ -35,6 +34,7 @@ import { z } from 'zod'
 import type { VideoResource } from '@coursebuilder/core/schemas'
 import {
 	Button,
+	DateTimePicker,
 	Dialog,
 	DialogContent,
 	DialogDescription,
