+++ packages/adapter-drizzle/src/lib/mysql/index.ts
@@ -59,6 +59,13 @@ import {
 	ContentResourceSchema,
 	type ContentResource,
 } from '@coursebuilder/core/schemas/content-resource-schema'
+import {
+	AdapterEventData,
+	AdapterEventSeriesData,
+	EventSchema,
+	EventSeriesResult,
+	EventSeriesSchema,
+} from '@coursebuilder/core/schemas/event-domain-schema'
 import { merchantAccountSchema } from '@coursebuilder/core/schemas/merchant-account-schema'
 import { merchantCustomerSchema } from '@coursebuilder/core/schemas/merchant-customer-schema'
 import {
@@ -349,6 +356,7 @@ export function mySqlDrizzleAdapter(
 		contentResource,
 		contentResourceResource,
 		contentResourceProduct,
+		contentResourceTag,
 		purchases: purchaseTable,
 		purchaseUserTransfer,
 		coupon,
@@ -2977,6 +2985,193 @@ export function mySqlDrizzleAdapter(
 			}
 
 			return subscriptionParsed.data
+		},
+		// Event methods
+		createEvent: async (data: AdapterEventData) => {
+			const hash = guid()
+			const newResourceId = `event-${hash}`
+
+			const newEvent = {
+				id: newResourceId,
+				organizationId: data.organizationId || null,
+				type: data.type,
+				fields: {
+					...data.fields,
+					slug:
+						data.fields.slug ||
+						`${data.fields.title.toLowerCase().replace(/\s+/g, '-')}-${hash}`,
+					state: data.fields.state || 'draft',
+					visibility: data.fields.visibility || 'public',
+				},
+				createdById: data.createdById,
+				createdByOrganizationMembershipId: null,
+				createdAt: new Date(),
+				updatedAt: new Date(),
+				deletedAt: null,
+				currentVersionId: null,
+			}
+
+			await client.insert(contentResource).values(newEvent)
+
+			const resource = await client.query.contentResource.findFirst({
+				where: eq(contentResource.id, newResourceId),
+			})
+
+			if (!resource) {
+				throw new Error('Failed to create event')
+			}
+
+			return ContentResourceSchema.parse(resource)
+		},
+		createEventSeries: async (
+			data: AdapterEventSeriesData,
+		): Promise<EventSeriesResult> => {
+			const result = await client.transaction(async (tx) => {
+				// Step 1: Create the event series resource
+				const eventSeriesHash = guid()
+				const eventSeriesResourceId = `event-series-${eventSeriesHash}`
+
+				const newEventSeries = {
+					id: eventSeriesResourceId,
+					organizationId: data.organizationId || null,
+					type: data.type,
+					fields: {
+						...data.eventSeries,
+						slug:
+							data.eventSeries.slug ||
+							`${data.eventSeries.title.toLowerCase().replace(/\s+/g, '-')}-${eventSeriesHash}`,
+						state: data.eventSeries.state || 'draft',
+						visibility: data.eventSeries.visibility || 'public',
+					},
+					createdById: data.createdById,
+					createdByOrganizationMembershipId: null,
+					createdAt: new Date(),
+					updatedAt: new Date(),
+					deletedAt: null,
+					currentVersionId: null,
+				}
+
+				await tx.insert(contentResource).values(newEventSeries)
+
+				// Fetch the created event series with relationships
+				const eventSeriesResource = await tx.query.contentResource.findFirst({
+					where: eq(contentResource.id, eventSeriesResourceId),
+					with: {
+						resources: {
+							with: {
+								resource: true,
+							},
+							orderBy: asc(contentResourceResource.position),
+						},
+						tags: {
+							with: {
+								tag: true,
+							},
+							orderBy: asc(contentResourceTag.position),
+						},
+						resourceProducts: {
+							with: {
+								product: {
+									with: {
+										price: true,
+									},
+								},
+							},
+						},
+					},
+				})
+
+				if (!eventSeriesResource) {
+					throw new Error('Failed to create event series')
+				}
+
+				// Step 2: Create child events and associate them
+				const createdChildEvents = []
+
+				for (let i = 0; i < data.childEvents.length; i++) {
+					const childEventInput = data.childEvents[i]
+
+					if (!childEventInput) {
+						throw new Error(`Child event input is required`)
+					}
+
+					// Create child event
+					const childEventHash = guid()
+					const childEventResourceId = `event-${childEventHash}`
+
+					const newChildEvent = {
+						id: childEventResourceId,
+						organizationId: data.organizationId || null,
+						type: childEventInput.type,
+						fields: {
+							...childEventInput.fields,
+							slug: `${childEventInput.fields.title.toLowerCase().replace(/\s+/g, '-')}-${childEventHash}`,
+							state: 'draft',
+							visibility: 'public',
+						},
+						createdById: data.createdById,
+						createdByOrganizationMembershipId: null,
+						createdAt: new Date(),
+						updatedAt: new Date(),
+						deletedAt: null,
+						currentVersionId: null,
+					}
+
+					await tx.insert(contentResource).values(newChildEvent)
+
+					// Fetch the created child event with relationships
+					const childEventResource = await tx.query.contentResource.findFirst({
+						where: eq(contentResource.id, childEventResourceId),
+						with: {
+							resources: {
+								with: {
+									resource: true,
+								},
+								orderBy: asc(contentResourceResource.position),
+							},
+							tags: {
+								with: {
+									tag: true,
+								},
+								orderBy: asc(contentResourceTag.position),
+							},
+							resourceProducts: {
+								with: {
+									product: {
+										with: {
+											price: true,
+										},
+									},
+								},
+							},
+						},
+					})
+
+					if (!childEventResource) {
+						throw new Error(`Failed to create child event ${i + 1}`)
+					}
+
+					// Associate child event with the event series
+					await tx.insert(contentResourceResource).values({
+						resourceOfId: eventSeriesResourceId,
+						resourceId: childEventResourceId,
+						position: i,
+						createdAt: new Date(),
+						updatedAt: new Date(),
+					})
+
+					createdChildEvents.push(childEventResource)
+				}
+
+				return {
+					eventSeries: EventSeriesSchema.parse(eventSeriesResource),
+					childEvents: createdChildEvents.map((event) =>
+						EventSchema.parse(event),
+					),
+				}
+			})
+
+			return result
 		},
 	}
