+++ packages/core/src/schemas/index.ts
@@ -13,3 +13,4 @@ export * from './upgradable-product-schema'
 export * from './merchant-charge-schema'
 export * from './merchant-price-schema'
 export * from './content-resource-schema'
+export * from './event-ui-schema'
